Fix invalidation failure when uninstalling a package
When one of apps which share a uid is uninstalled, the cached instance
of LocalCallingIdentity is not invalidated. The cached instance can get
wrong value for permissions like W_E_S or M_E_S if
LocalCallingIdentity#packageNameUnchecked is package name of the
uninstalled app because package name of the uninstalled app is used to
see if W_E_S or M_E_S is granted or not.

Bug: 194111021
Test: atest MediaProviderTests
Change-Id: Ie4b07df5cfc41fbc497244da6e0f9300789a3294