Adding verifications for CloudMediaProvider Apis.
This will help in debugging and keeping a check on latencies for the API
executions time and also will verify that the contract is being
respected.

Bug: 285895977

Set up:
Ensure cloud is enabled and a valid CloudMediaProvider is present.
Run the following adb command:
```
adb shell setprop log.tag.CloudMediaProvider VERBOSE
```

Test: through logcat 'adb logcat | grep CMP_verifications_{$authority}'

Change-Id: I5b4908d89d7580c36dc76c1143110b9340f22eba