Don't allocate a buffer if not related to xmp
It's reasonable not to allocate a buffer when the box type is 'uuid' and
value of the uuid is not 'be7acfcb-97a9-42e8-9c71-999491e3afac' to
prevent OOME from occurring.

Also skip XMP boxes that have data > 1 MB. XMP data shouldn't exceed 64KB.

Additionally, keep in memory only boxes with types that we are interested in to prevent from OOME in case of files having too many boxes that most of them we don't make use of.

Bug: 306057435
Flag: EXEMPT bug fix
Test: atest MediaProviderTests
Change-Id: Ib7d1f828a171ab9fd10696a2fe609e279e4eb00e